fix(smol_stuff): Smol stuff
