+++ src/types/position.rs
@@ -1,11 +1,9 @@
-use anyhow::{anyhow, Context, Ok, Result};
+use anyhow::{anyhow, Ok, Result};
 use apibara_core::starknet::v1alpha2::FieldElement;
 use bigdecimal::num_bigint::BigInt;
 use bigdecimal::{BigDecimal, FromPrimitive};
-use cainome::cairo_serde::{ContractAddress, U256};
 use colored::Colorize;
 use serde::{Deserialize, Serialize};
-use serde_json::Value;
 use starknet::core::types::{BlockId, BlockTag, FunctionCall};
 use starknet::core::types::{Call, Felt};
 use starknet::providers::jsonrpc::HttpTransport;
@@ -16,9 +14,7 @@ use std::hash::{Hash, Hasher};
 use std::sync::Arc;
 use tokio::sync::RwLock;
 
-use crate::bindings::liquidate::{
-    Liquidate, LiquidateParams, PoolKey, RouteNode, Swap, TokenAmount, I129,
-};
+use crate::bindings::liquidate::{Liquidate, LiquidateParams, RouteNode, Swap, TokenAmount, I129};
 
 use crate::config::{Config, LiquidationMode, LIQUIDATION_CONFIG_SELECTOR};
 use crate::services::oracle::LatestOraclePrices;
@@ -26,12 +22,13 @@ use crate::storages::Storage;
 use crate::utils::apply_overhead;
 use crate::utils::constants::VESU_RESPONSE_DECIMALS;
 use crate::utils::conversions::big_decimal_to_felt;
+use crate::utils::ekubo::get_ekubo_route;
 use crate::{types::asset::Asset, utils::conversions::apibara_field_as_felt};
 
 use super::account::StarknetAccount;
 
 /// Threshold for which we consider a position almost liquidable.
-const ALMOST_LIQUIDABLE_THRESHOLD: f64 = 0.03;
+const ALMOST_LIQUIDABLE_THRESHOLD: f64 = 0.02;
 
 /// Thread-safe wrapper around the positions.
 /// PositionsMap is a map between position position_key <=> position.
@@ -250,84 +247,6 @@ impl Position {
         hasher.finish()
     }
 
-    pub async fn get_ekubo_route(
-        amount_as_string: String,
-        from_token: String,
-        to_token: String,
-    ) -> Result<Vec<RouteNode>> {
-        let ekubo_api_endpoint = format!(
-            "https://mainnet-api.ekubo.org/quote/{amount_as_string}/{from_token}/{to_token}"
-        );
-        let http_client = reqwest::Client::new();
-
-        let response = http_client.get(ekubo_api_endpoint).send().await?;
-
-        if !response.status().is_success() {
-            anyhow::bail!("API request failed with status: {}", response.status());
-        }
-
-        let response_text = response.text().await?;
-
-        let json_value: Value = serde_json::from_str(&response_text)?;
-
-        // We have to deserialize by hand into a Vec of [RouteNode].
-        // TODO: Make this cleaner!
-        let route = json_value["route"]
-            .as_array()
-            .context("'route' is not an array")?
-            .iter()
-            .map(|node| {
-                let pool_key = &node["pool_key"];
-                Ok(RouteNode {
-                    pool_key: PoolKey {
-                        token0: ContractAddress(Felt::from_hex(
-                            pool_key["token0"]
-                                .as_str()
-                                .context("token0 is not a string")?,
-                        )?),
-                        token1: ContractAddress(Felt::from_hex(
-                            pool_key["token1"]
-                                .as_str()
-                                .context("token1 is not a string")?,
-                        )?),
-                        fee: u128::from_str_radix(
-                            pool_key["fee"]
-                                .as_str()
-                                .context("fee is not a string")?
-                                .trim_start_matches("0x"),
-                            16,
-                        )
-                        .context("Failed to parse fee as u128")?,
-                        tick_spacing: pool_key["tick_spacing"]
-                            .as_u64()
-                            .context("tick_spacing is not a u64")?
-                            as u128,
-                        extension: ContractAddress(Felt::from_hex(
-                            pool_key["extension"]
-                                .as_str()
-                                .context("extension is not a string")?,
-                        )?),
-                    },
-                    sqrt_ratio_limit: U256::from_bytes_be(
-                        &Felt::from_hex(
-                            node["sqrt_ratio_limit"]
-                                .as_str()
-                                .context("sqrt_ratio_limit is not a string")?,
-                        )
-                        .unwrap()
-                        .to_bytes_be(),
-                    ),
-                    skip_ahead: node["skip_ahead"]
-                        .as_u64()
-                        .context("skip_ahead is not a u64")?
-                        as u128,
-                })
-            })
-            .collect::<Result<Vec<RouteNode>>>()?;
-
-        Ok(route)
-    }
-
     /// Returns the TX necessary to liquidate this position (approve + liquidate).
     // See: https://github.com/vesuxyz/vesu-v1/blob/a2a59936988fcb51bc85f0eeaba9b87cf3777c49/src/singleton.cairo#L1624
     pub async fn get_liquidation_txs(
@@ -349,19 +268,23 @@ impl Position {
         };
 
         // As mentionned before the route is inverted for precision purpose
-        let liquidate_route: Vec<RouteNode> = Position::get_ekubo_route(
+        let liquidate_route: Vec<RouteNode> = get_ekubo_route(
             String::from("10"), // TODO: Investigate the behavior of this value with the Vesu Liquidate contract
             self.debt.name.clone(),
             self.collateral.name.clone(),
         )
         .await?;
 
-        let withdraw_route: Vec<RouteNode> = Position::get_ekubo_route(
-            String::from("10"), // TODO: Investigate the behavior of this value with the Vesu Liquidate contract
-            self.debt.name.clone(),
-            String::from("usdc"),
-        )
-        .await?;
+        let withdraw_route: Vec<RouteNode> = if self.debt.name == "USDC" {
+            vec![]
+        } else {
+            get_ekubo_route(
+                String::from("10"), // TODO: Investigate the behavior of this value with the Vesu Liquidate contract
+                self.debt.name.clone(),
+                String::from("USDC"),
+            )
+            .await?
+        };
 
         let liquidate_contract = Liquidate::new(liquidate_contract, account.0.clone());
